# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fnashha Home Services Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create service (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

    patch:
      operationId: updateService
      tags: [services]
      summary: Update service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PUBLIC SEO LANDING PAGES ───────────────────────────────────────────────
...
  /governorates:
    get:
      operationId: listGovernorates
      tags: [governorates]
      summary: List governorates
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Governorate"
    post:
      operationId: createGovernorate
      tags: [governorates]
      summary: Create governorate (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernorateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Governorate"
  /governorates/{id}:
    patch:
      operationId: updateGovernorate
      tags: [governorates]
      summary: Update governorate (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernorateUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Governorate"
    delete:
      operationId: deleteGovernorate
      tags: [governorates]
      summary: Delete governorate (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── AREAS ───────────────────────────────────────────────────────────────────
  /areas:
    get:
      operationId: listAreas
      tags: [areas]
      summary: List areas
      parameters:
        - name: governorateId
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Area"
    post:
      operationId: createArea
      tags: [areas]
      summary: Create area (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AreaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
  /areas/{id}:
    patch:
      operationId: updateArea
      tags: [areas]
      summary: Update area (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AreaUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
    delete:
      operationId: deleteArea
      tags: [areas]
      summary: Delete area (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── REQUESTS ────────────────────────────────────────────────────────────────
...
  /points/transactions:
    get:
      operationId: listPointTransactions
      tags: [points]
      summary: List point transactions
      parameters:
        - name: technicianId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PointTransaction"
  /points/add:
    post:
      operationId: addPoints
      tags: [points]
      summary: Add points to technician (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointsBalance"
  /points/deduct:
    post:
      operationId: deductPoints
      tags: [points]
      summary: Deduct points from technician (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointsBalance"

  # ─── COMMISSIONS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, name, isActive, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
        displayOrder:
          type: integer
...
    ServiceInput:
      type: object
      required: [name, nameAr]
      properties:
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
        image:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
        image:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer

    # ─── GOVERNORATE ───────────────────────────────────────────────────────────
    Governorate:
      type: object
      required: [id, name, nameAr, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean
    GovernorateInput:
      type: object
      required: [name, nameAr]
      properties:
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean
    GovernorateUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean

    # ─── AREA ──────────────────────────────────────────────────────────────────
    Area:
      type: object
      required: [id, name, nameAr, governorateId, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        governorateId:
          type: integer
        governorate:
          $ref: "#/components/schemas/Governorate"
        isActive:
          type: boolean
    AreaInput:
      type: object
      required: [name, nameAr, governorateId]
      properties:
        name:
          type: string
        nameAr:
          type: string
        governorateId:
          type: integer
        isActive:
          type: boolean
    AreaUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean

    # ─── REQUEST ───────────────────────────────────────────────────────────────
...
    PointsBalance:
      type: object
      required: [technicianId, balance]
      properties:
        technicianId:
          type: integer
        balance:
          type: integer
    PointTransaction:
      type: object
      required: [id, technicianId, amount, type, description, balanceAfter, createdAt]
      properties:
        id:
          type: integer
        technicianId:
          type: integer
        amount:
          type: integer
        type:
          type: string
          enum: [credit, debit, commission]
        description:
          type: string
        balanceAfter:
          type: integer
        requestId:
          type: ["integer", "null"]
        createdAt:
          type: string
    PointsInput:
      type: object
      required: [technicianId, amount, description]
      properties:
        technicianId:
          type: integer
        amount:
          type: integer
        description:
          type: string

    # ─── COMMISSION ────────────────────────────────────────────────────────────
...